src/index.ts
import {
  Configuration,
  ServiceConfiguration,
  RabbitConfiguration,
  ApiConfiguration,
  serviceConf,
  apiConf,
  rabbitConf,
  configuration,
} from './utils/ServiceConf';

import {
  Base,
  HandlerBase,
  EventHandlerBase,
  TaskHandlerBase,
  RPCHandlerBase,
  RouteHandlerBase,
} from './base';
import ApiRoute from './base/RouteHandlerBase'; // TODO: remove duplicated renamed interface

import {
  Express,
  NextFunction,
  Request,
  Response,
} from './utils/API';

import Logger, { LoggerMode } from './utils/Logger';

import ServiceBase, {
  ServiceResources,
} from './ServiceBase';

export {
  ApiRoute, // TODO: remove duplicated renamed interface
  RouteHandlerBase,
  Configuration,
  ServiceResources,
  ServiceBase,
  Base,
  HandlerBase,
  EventHandlerBase,
  TaskHandlerBase,
  RPCHandlerBase,
  Express,
  NextFunction,
  Response,
  Request,
  Logger,
  LoggerMode,
  ServiceConfiguration,
  RabbitConfiguration,
  ApiConfiguration,
  serviceConf,
  apiConf,
  rabbitConf,
  configuration,
};
